Refactor tests to improve clarity and ensure proper handling of return values
- Updated assertions in various test files to use static_cast<void>() for ignored return values, enhancing code clarity and intent.
- Simplified test logic by directly asserting expected outcomes without unnecessary intermediate variables.
- Improved error handling in tests by ensuring exceptions are caught and handled appropriately.
- Enhanced readability by using consistent naming conventions and formatting across test cases.
- Ensured that all tests maintain their original functionality while improving maintainability and readability.

+++ src/llm/inference_engine_enhanced.cpp
@@ -1627,10 +1627,10 @@ void InferenceEngineEnhanced::recordBatchCompletion(size_t batch_size) {
     
     // Update moving average
     if (stats_.avg_batch_size == 0.0) {
-        stats_.avg_batch_size = batch_size;
+        stats_.avg_batch_size = static_cast<double>(batch_size);
     } else {
         stats_.avg_batch_size = 
-            0.95 * stats_.avg_batch_size + 0.05 * batch_size;
+            0.95 * stats_.avg_batch_size + 0.05 * static_cast<double>(batch_size);
     }
     
     if (batch_size > stats_.max_batch_size_seen) {
