+++ src/rag/llm_judge_client.cpp
@@ -236,8 +236,8 @@ std::string LLMJudgeClient::evaluate(const std::string& prompt) {
         llm::InferenceEngineEnhanced::EnhancedInferenceRequest request;
         request.base_request.prompt = prompt;
         request.base_request.max_tokens = impl_->config.max_tokens;
-        request.base_request.temperature = impl_->config.temperature;
-        request.base_request.top_p = 0.95;
+        request.base_request.temperature = static_cast<float>(impl_->config.temperature);
+        request.base_request.top_p = 0.95f;
         request.base_request.stop_sequences = impl_->config.stop_sequences;
         
         request.priority = impl_->config.priority;
@@ -290,8 +290,8 @@ std::vector<std::string> LLMJudgeClient::evaluateBatch(
             llm::InferenceEngineEnhanced::EnhancedInferenceRequest request;
             request.base_request.prompt = prompt;
             request.base_request.max_tokens = impl_->config.max_tokens;
-            request.base_request.temperature = impl_->config.temperature;
-            request.base_request.top_p = 0.95;
+            request.base_request.temperature = static_cast<float>(impl_->config.temperature);
+            request.base_request.top_p = 0.95f;
             request.base_request.stop_sequences = impl_->config.stop_sequences;
             
             request.priority = impl_->config.priority;
